Fix random numbers distribution
Added getRandomInt function to get an even distribution
of random numbers between 0, 1 and 2 with the max technique.
This makes the getComputerNumber function a true random
choice between 3 values (before the distribution was in
favor of rock which had 0-33 = 34 integers, including the zero,
against paper and scissors which only had 33 integers to
chose from respectively).