Fix playRound() and edit getHumanChoice()
Added toLowerCase() method inside the getHumanChoice
function to get better message on console and for not
having to do it later.
Edited the playRound function so that it works as intended.
Initially thought of using only a switch statement, I ended
up using both if ... else if statements and a switch with
two cases per condition.
Note to self: look for a more concise method.